LayoutTests/media/W3C/audio/canPlayType/canPlayType_bogus_type.html

<!doctype html>
<html>
 <head>
  <title>audio.canPlayType() - bogus format</title>
  <script src="../../w3cwrapper.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-navigator-canplaytype">spec reference</a></p>
  <audio id="a">
  </audio>
  <div id="log"></div>
  <script>
test(function() {
 assert_equals(
  document.getElementById("a").canPlayType("application/marks-fantasmagorical-format"),
  "",
  "audioElement.canPlayType() should return the empty string on a bogus format");
});
  </script>
 </body>
</html>
