LayoutTests/media/W3C/audio/canPlayType/canPlayType_bogus_type-expected.txt

audio.canPlayType() - bogus format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "" is ""

TEST COMPLETE
spec reference
